feat(cli): reorganize internal structure and add exports for all utilities
