+++ packages/cli/src/commands/deprecated/call-system.ts
@@ -1,8 +1,8 @@
 import { defaultAbiCoder as abi } from "ethers/lib/utils.js";
 import path from "path";
 import type { CommandModule } from "yargs";
-import { execLog } from "../utils/index.js";
-import { getTestDirectory } from "../utils/forgeConfig.js";
+import { execLog } from "../../utils/deprecated/index.js";
+import { getTestDirectory } from "../../utils/forgeConfig.js";
 
 type Options = {
   rpc?: string;
