+++ packages/cli/src/index.ts
@@ -1,1 +1,7 @@
+import { loadStoreConfig, parseStoreConfig } from "./config/loadStoreConfig.js";
+import { renderTablesFromConfig } from "./render-table/renderTablesFromConfig.js";
+import { renderTable } from "./render-table/renderTable.js";
+
 export type { StoreUserConfig, StoreConfig } from "./config/loadStoreConfig.js";
+
+export { loadStoreConfig, parseStoreConfig, renderTablesFromConfig, renderTable };
