+++ packages/cli/src/utils/deprecated/index.ts
@@ -1,0 +1,8 @@
+export * from "./exec.js";
+export * from "./ids.js";
+export * from "./codegen.js";
+export * from "./deploy.js";
+export * from "./hsr.js";
+export * from "./findLog.js";
+export * from "./typegen.js";
+export * from "./build.js";
