+++ packages/cli/src/utils/deprecated/typegen.ts
@@ -4,7 +4,7 @@ import { extractIdFromFile } from "./ids.js";
 import { rmSync, writeFileSync } from "fs";
 import path from "path";
 import { filterAbi, forgeBuild } from "./build.js";
-import { getOutDirectory, getSrcDirectory } from "./forgeConfig.js";
+import { getOutDirectory, getSrcDirectory } from "../forgeConfig.js";
 import { systemsDir } from "./constants.js";
 
 export async function generateAbiTypes(
