+++ packages/cli/tsup.config.js
@@ -1,14 +1,21 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/index.ts", "src/mud.ts"],
+  entry: [
+    "src/index.ts",
+    "src/mud.ts",
+    "src/config/index.ts",
+    "src/render-table/index.ts",
+    "src/utils/index.ts",
+    "src/utils/deprecated/index.ts",
+  ],
   outDir: "dist",
   platform: "node",
   format: ["esm"],
   sourcemap: false,
   clean: true,
   bundle: true,
-  dts: "src/index.ts",
+  dts: true,
   // both inlined imports are hacks to make esm+esbuild work
   // @latticexyz/services: esm imports directly from ts files don't work properly from transpiled js
   // @latticexyz/solecs: esm requires import assertions for json, but esbuild doesn't support those (they just get stripped)
